+++ react/AutocompleteBlock.tsx
@@ -273,10 +273,12 @@ const AutocompleteBlock: FunctionComponent<any & WrappedComponentProps> = ({
   }
 
   const CSS_HANDLES = [
+    'autoCompleteBlock',
     'skuSelection',
     'productThumb',
     'productTitle',
     'productSku',
+    'skuSelected',
     'productLabel',
     'inputQuantity',
     'buttonAdd',
@@ -290,7 +292,7 @@ const AutocompleteBlock: FunctionComponent<any & WrappedComponentProps> = ({
   const handles = useCssHandles(CSS_HANDLES)
 
   return (
-    <div>
+    <div className={`${handles.autoCompleteBlock}`}>
       {!componentOnly && (
         <div className={`${handles.textContainer} w-third-l w-100-ns fl-l`}>
           <h2
@@ -340,7 +342,10 @@ const AutocompleteBlock: FunctionComponent<any & WrappedComponentProps> = ({
                             return (
                               <span
                                 key={item.itemId}
-                                className={`mr4 ${handles.skuSelection}`}
+                                className={`mr4 ${handles.skuSelection} ${
+                                  item.itemId === selectedItem.value &&
+                                  handles.skuSelected
+                                }`}
                               >
                                 <Tag
                                   size="small"